Expose block in TinderBox to ease project creation
The Syphon framework is copied to the build product via the buildCopy XML element, but requires f7a0854a53 on the Cinder dev branch to recognize the destination setting of 'Frameworks' instead of the default 'Executables'. Users of earlier Cinder/TinderBox versions will need to manually change the destination in the 'Copy Files' phase in the 'Build Phases' section for their application target.